extract functionality of delete button into new component
