+++ src/components/DisplaySingleRecipe.jsx
@@ -1,12 +1,12 @@
 import React, { useState, useEffect } from "react";
-import { useParams } from "react-router-dom";
+import { useParams, useNavigate } from "react-router-dom";
 import Recipes from "../modules/recipes";
 import CircularProgress from "@mui/material/CircularProgress";
 import { Button } from "@mui/material";
-import { useNavigate } from "react-router-dom";
 import DisplayIngredient from "./DisplayIngredient";
 import DisplayInstruction from "./DisplayInstruction";
 import Headline from "./Headline";
+import SingleRecipeDelete from "./SingleRecipeDelete";
 
 const DisplaySingleRecipe = () => {
   const [recipe, setRecipe] = useState();
@@ -21,12 +21,6 @@ const DisplaySingleRecipe = () => {
     fetchRecipe();
   }, [id]);
 
-  const handleDelete = async () => {
-    const response = await Recipes.delete(id);
-    const message = response.data.message;
-    navigate("/", { state: { message: message } });
-  };
-
   if (recipe === undefined) {
     return <CircularProgress />;
   } else {
@@ -36,14 +30,8 @@ const DisplaySingleRecipe = () => {
 
     return (
       <>
-        <Button
-          onClick={handleDelete}
-          data-cy="delete-btn"
-          color="secondary"
-          variant="contained"
-        >
-          Delete this recipe
-        </Button>
+        <Headline viewHeadline={recipe.title} />
+        <SingleRecipeDelete />
         <Button
           data-cy="edit-recipe-btn"
           color="secondary"
@@ -52,7 +40,6 @@ const DisplaySingleRecipe = () => {
         >
           Edit
         </Button>
-        <Headline viewHeadline={recipe.title} />
         <DisplayIngredient ingredients={recipe.ingredients} />
         <ol>{instructions}</ol>
       </>
